ai_service/requirements.txt
kafka-python==2.0.2
scikit-learn==1.4.2
numpy==1.26.4
joblib==1.4.2
python-dotenv==1.0.1
